refactor: MessageResolver 클래스 OutputView로 통합
